requirements.txt
python-dotenv
pyTelegramBotAPI
google-generativeai